ai/langchain_api/requirements.txt
fastapi
uvicorn
pydantic
langserve[all]
langchain
langchain-community
llama-cpp-python
python-dotenv